+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), following [Se
 
 ## [Unreleased]
 
+- added: metadata includes now `annotations` and `chartApiVersion` fields
+
 ### added
 
 - Initial commit